+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1263,5 +1263,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-05T20:13:44.106Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-17T07:48:11.095Z")
     r"""The date and time the resource was last updated"""