+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K4DR766MPD6TGW5TVY2HSRG3"
+        "01K5BAAP2RT6192PPZ1WE61YSF"
     )
     r"""The id of the resource"""