+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -18,12 +18,12 @@ ModelType = Literal[
     "chat",
     "completion",
     "embedding",
-    "vision",
     "image",
     "tts",
     "stt",
     "rerank",
     "moderations",
+    "vision",
 ]
 r"""The modality of the model"""
 
@@ -702,12 +702,12 @@ CreatePromptModelType = Literal[
     "chat",
     "completion",
     "embedding",
-    "vision",
     "image",
     "tts",
     "stt",
     "rerank",
     "moderations",
+    "vision",
 ]
 r"""The modality of the model"""