+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -130,12 +130,12 @@ DeploymentsModelType = Literal[
     "chat",
     "completion",
     "embedding",
-    "vision",
     "image",
     "tts",
     "stt",
     "rerank",
     "moderations",
+    "vision",
 ]
 r"""The modality of the model"""