+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -114,8 +114,6 @@ class GetBudgetResponseBodyTypedDict(TypedDict):
     r"""Whether this budget configuration is currently active"""
     contact_id: NotRequired[str]
     r"""Contact external identifier (present when type is \"contact\")"""
-    api_key_id: NotRequired[str]
-    r"""API Key identifier (present when type is \"api_key\")"""
     consumption: NotRequired[GetBudgetConsumptionTypedDict]
     created: NotRequired[datetime]
     r"""The date and time the resource was created"""
@@ -141,13 +139,10 @@ class GetBudgetResponseBody(BaseModel):
     contact_id: Optional[str] = None
     r"""Contact external identifier (present when type is \"contact\")"""
 
-    api_key_id: Optional[str] = None
-    r"""API Key identifier (present when type is \"api_key\")"""
-
     consumption: Optional[GetBudgetConsumption] = None
 
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-05T20:13:44.106Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-17T07:48:11.095Z")
     r"""The date and time the resource was last updated"""