+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -33,12 +33,12 @@ GetOnePromptModelType = Literal[
     "chat",
     "completion",
     "embedding",
-    "vision",
     "image",
     "tts",
     "stt",
     "rerank",
     "moderations",
+    "vision",
 ]
 r"""The modality of the model"""