+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -55,12 +55,12 @@ ListPromptVersionsModelType = Literal[
     "chat",
     "completion",
     "embedding",
-    "vision",
     "image",
     "tts",
     "stt",
     "rerank",
     "moderations",
+    "vision",
 ]
 r"""The modality of the model"""