+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K4DR766PJGVJQ0M0WSCMP720"
+        "01K5BAAP2SM8DXGQM3TJQ34GB0"
     )
     r"""The id of the resource"""